dashboard/requirements.txt
flask>=3.0
apscheduler>=3.10
requests>=2.31
pillow>=10.0
